chore: move data models
