+++ src/CoreLib/AntiAmbrose/AuthorityRequester.cs
@@ -4,6 +4,7 @@
  */
 
 using Imlight.Common;
+using Imlight.CoreLib.WizardData.Models.Misc;
 using Imlight.CoreLib.WizardData.Models.Player;
 
 namespace Imlight.CoreLib.AntiAmbrose;
