+++ src/CoreLib/Game/Commands/Protocols/CommandAccountProtocol.cs
@@ -6,6 +6,7 @@
 using Imlight.CoreLib.AntiAmbrose;
 using Imlight.CoreLib.WizardData;
 using Imlight.CoreLib.WizardData.Implementations;
+using Imlight.CoreLib.WizardData.Models.Misc;
 using Imlight.CoreLib.WizardData.Models.Player;
 using System;
 using System.Text;
