+++ src/CoreLib/Game/Commands/Protocols/CommandPunishmentProtocol.cs
@@ -4,6 +4,7 @@
  */
 
 using Imlight.CoreLib.Shared.Packets;
+using Imlight.CoreLib.WizardData.Models.Misc;
 using Imlight.CoreLib.WizardData.Models.Player;
 using System;
 using System.Text;
