+++ src/CoreLib/WizardData/Collections/AccountCollection.cs
@@ -6,6 +6,7 @@
 using System.Linq;
 using Imlight.CoreLib.WizardData.Collections;
 using Imlight.CoreLib.WizardData.Databases;
+using Imlight.CoreLib.WizardData.Models.Misc;
 using Imlight.CoreLib.WizardData.Models.Player;
 using Raven.Client.Documents;
 using static Imlight.Common.Caches.TypeCache;
