+++ src/CoreLib/WizardData/Collections/InfractionCollection.cs
@@ -6,7 +6,7 @@
 using System;
 using System.Linq;
 using Imlight.CoreLib.WizardData.Databases;
-using Imlight.CoreLib.WizardData.Models.Player;
+using Imlight.CoreLib.WizardData.Models.Misc;
 using Raven.Client.Documents;
 
 namespace Imlight.CoreLib.WizardData.Collections;
